Extract movie and movielist into own components.
