+++ src/components/App/components/MovieList/MovieList.js
@@ -1,0 +1,8 @@
+import {Movie} from './components/Movie/Movie';
+
+export function MovieList({movieResults, nominees, onNominate}) {
+    const listItems = movieResults.map((movieResult) =>
+        <li key={JSON.stringify(movieResult.imdbID)}><Movie movie={movieResult} nominees={nominees} onNominate={onNominate} /></li>
+    );
+    return <ul>{listItems}</ul>;
+}
